Use only named packages from extra registries
This change prevents errors from RegistryEntry::get_path_to_version()
when the default registry doesn't have a port for a name which is
part of some extra registry but not listed that registry's "packages"
configuration field.
This is consistent with RegistrySet::registry_for_port() which only
considers the the named packages from the extra registries.