Avoid repeated binary cache server operations.
We recently have run into problems in our normal CI infrastructure when we increased the size of the VMs where all VMs tried to fetch the same nonexistent entries over and over and caused failures. To alievate that problem, this changes binary caching to not issue server I/Os for status that vcpkg already knows.

To that end, this is a substantial redesign of the IBinaryProvider mechanism to give this global knowledge of which things have been restored from cache, such that MergedBinaryProvider is replaced with class BinaryCache, and the functions that were 'masaging' the interface to IBinaryProvider are put in that type. Moreover, NullBinaryProvder is an empty binary cache.

Incidential fixes:
* Different commands were inconsistent in how they configured the binary cache, in particular manifest mode and normal mode installs treated --download-only differently.
* The different IBinaryProviders that make a temp zip file all use the same zip file name.