Fixed issue with output from CMake 3.21.0
When calling CMake to obtain triplet information a regression change from CMake 3.20.5 to 3.21.0-rc1
causes the tool to exit due to an exception.

This is due to the response containing a mixture of 'n' and  '\r\n' line endings.

After splitting the output on line feeds ('\n') each line is then trimmed of whitespace to remove
any carriage returns ('\r') or trailing whitespace. This could be fixed in CMake but this
change makes it robust.

This commit in CMake caused the fault: https://github.com/Kitware/CMake/commit/0a0a0f8a744e2bfa35cdbc90db6e4e23adadd59b
It is unlikely to change as they have added a new helper file for writing the terminal output