Preserve RAII for file locking through virtual dispatch. (#152)
* Preserve RAII for file locking through virtual dispatch.

This change was extracted from adding the POSIX backend for filesystem to be easier to review. It moves everything related to file locking into files.cpp, and eliminates ifdefs in files.h for platform selection. It uses std::function to unlock the lock instead. Notably, this is the same number of virtual dispatches as before, and eliminates a theoretical lifetime problem where the Filesystem* did not outlive the ExclusiveFileLock type.

* Update include/vcpkg/base/files.h

Co-authored-by: Billy O'Neal <bion@microsoft.com>

Co-authored-by: nicole mazzuca <83086508+strega-nil-ms@users.noreply.github.com>