Fix copy_symlink on old libstdc++.
`copy_symlink` had assumptions that `std::string` implements the small string optimization, and thus we would never call `readlink` with a buffer length of 0. Unfortunately, the old libstdc++ we want to use for old Linux uses the COW form where the initial capacity is 0, which caused `readlink` to fail with `EINVAL`.

This change makes us always start with `PATH_MAX`.