[tool cache] Add option to require exact versions for abi relevant tools
If a developer in a team has a different version of a abi relevant tool, the developer has
to rebuild every package because the developer only gets cache misses in the binary cache.
Add an option to force vcpkg to use the same version of a abi relevant tool so that you don't
get binary cache misses because of different abi relevant tools.