Fix CI crash due to missing package ABIs
PR #298 introduced a regression during ABI calculation for packages
that do not have features. It removed the implicit entry "core"
from InstallPlanAction::feature_list, which meant that packages
without features would end up with an empty ABI tag for their
features list.

We consider an empty ABI tag in the features list to indicate that
a value could not be calculated and therefore that the package
cannot be cached (no ABI). This then resulted in a crash during CI
which assumes that all packages have ABIs.