vcpkg install - cmake version check should NOT reject valid response.
As concisely stated by @BRKVings - There are two types of output:
- cmake3 version x.x.x
- cmake version x.x.x
But the current code can't handle the first case.

This minor annoyance, just turned into a hard-blocker for me. I am proposing a change, which works in a minimal-repro program in Visual Studio 2022. I'm not familiar with your code-base. Could someone please test and promote, or propose an alternative that is inline VCPkg style guidelines?