Lock VCPKG_BASE_VERSION in the first job for use in subsequent jobs.
The "2022-02-01" release is broken because the CE build thought the date was 2022-01-31, while the vcpkg.exe build thought it was 2022-02-01. See https://devdiv.visualstudio.com/DefaultCollection/DevDiv/_build/results?buildId=5694458&view=results

The pipeline.startTime variable appears to be reevaluated each time a job starts, and refers to the *job* start time, not the overall pipeline start time, so when a build crosses 4PM pacific time it produced broken output.

This change gets the first arch_independent stage to write the value it used, and this is fed into the remaining stages.