[format] remove disable 6385
Some old code did `const char* top = data::digits[exp / 100]`;
in that case, `exp < 10000 => exp / 100 < 100`, and thus that was not broken.
However, in new code they now do:

`const char* top = digits2(to_unsigned(exp / 100))`,
and thus this warning is no longer hit.

Since this is a useful warning, we don't want to disable it,
and thus we are re-enabling it.