Add --create-dirs to curl call in download_files_inner
Without this option, curl does not create intermediate directories
and thus trying to retrieve binary cache from az blob storage
fails if the `buildtrees` and its various subdirectories for ports
are not already present.

The alternative is for vcpkg to create the intermediate dirs
explicitly, but we already use the option in `try_donwload_file`,
so there is no reason not to use it here as well.