Fix localization irregularities. (#553)
* Delete LocalizedString::appendnl.

I believe appendnl hails from a time when we thought we were going to be much more rigid about manipulation of LocalizedStrings as tokens. There are localized strings which themselves contain newlines. There are also formatted data strings, such as error message blocks join'd with \n, that we need to insert into localized messages. We also aren't likely to ever want to change our newlines to \r\n or something like that :)

To that end, I think both original reasons for this to exist are gone, and it should go away.

I think we should go all the way to removal rather than only removing the guidelines about it, because I think we, and contributors, will be confused if it exists and is relatively meaningless in the same way that people are confused (and use) std::endl today. The fact is that newlines aren't special in our LocalizedString structures, and we should be honest about that in our design.

Contrast/note: append_indent must stay, and we must add a plain std::string version of that, to ensure consistency in how we handle indenting, as this is likely to be actually ambiguous (as opposed to newlines which are always and forever \n in vcpkg).

* Improve formatting of generated comments, and deduplicate code that generates comments.

* Remove extra newlines from localized messages that are being fed into println or similar components that already insert newlines.

* Fix indents in "VcpkgHasCrashed" message. Also direct people to GitHub rather than email.

* Also fix UnsupportedToolchain message.

* Also fix VersionSpecMismatch message.

* Deduplicate all Error:s in localized messages.

* Add enforcement for not including indenting, error:, or warning: in localized messages, and fix remaining edge case occurrences. Also removes allow-incorrect-comments from generate-message-map considering that we control all inputs here and extending this to block these other lints seemed unnecessary.

* Also block trailing newlines.

* Fix end to end test.

* Rename print_error and print_warning to println_error and println_warning, respectively.