Don't attempt to default-construct a reference.
Currently, Expected<T> has a bug for references, in that it is default constructible, but references shouldn't be default constructible. (This is hidden because we store a pointer internally).

In VersionedPackageGraph::require_port_version, an ExpectedS<T&> is default constructed, which is in the "impossible" state of neither having a stored T (since references can't be default constructed), but also not being assigned an error state.

This change hoists the only place where the ExpectedS could contain an error into an early-return, allowing a plain pointer to be used, and avoiding constructing the impossible thing.

(The bug in Expected that allowed this to happen is fixed in https://github.com/microsoft/vcpkg-tool/pull/511 , but this change was tricky enough I thought extracting it for review alone made sense)

Reasoning, using "after" line numbers:

1557-1567: This if block could never have stored an error state in the expected, since it is only entered if there is an overlay control file that will be the result. Therefore, we will never add an error to m_errors from here.
1578: get_control_file can return an expected in the error state, so we have to check if it's an error and, if so, add it to m_errors.
1587: At this point, we know p_scfl is not nullptr since both if branches set it on 1556 or 1578.