Load overlay ports in the correct order.
In OverlayProviderImpl::load_all_control_files, overlay directories are consulted in consecutive order. Each consulted directory overwrites any overlays that were visited previously. This means that given multiple overlays, the last one wins.

However, in OverlayProviderImpl::load_port, the overlay directories are still consulted in consecutive order, but the first one is chosen. This means that the result will be different than load_all_control_files if multiple overlays contain the same port.

Based on e2e tests, the intent appears to be that the first overlay wins, so we need to do load_all_control_files in reverse order.