[artifacts] Make various roots consistent between vcpkg and artifacts.
This is the first chunk of work for https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1494960 "[vcpkg ce] vcpkg-ce does not find vcpkg.json to find vcpkg-configuration.json"

In keeping with https://github.com/microsoft/vcpkg-tool/pull/601#issuecomment-1165742681 , this adds new parameters to ce for each of the directories in VcpkgPaths that ce cares about, and wires the C++ code to add parameters filling in those values as needed. TypeScript world keeps fallbacks in place but the vast majority of the time we expect the wrapping exe to be used.

Other changes:
* vcpkg-configuration.json is now the only accepted profile name.
* vcpkg-configuration.global.json is now replaced with $VCPKG_ROOT/vcpkg-configuration.json, which is consistent with what vcpkg does for manifest-less / classic mode operation.
* In several places, "cache" is renamed to "downloads" for consistency with the existing naming elsewhere, and also to avoid confusion with "binary caching" or "artifact caching".
* Implemented "deactivate".
* Updated the scripts tree sha to one that includes the node 16.15.1 update ( https://github.com/microsoft/vcpkg/commit/032d9d0820db290ce9ff644dabfdf564343013ae ) which drops the need for the `--harmony` switch.
* Changes handling of X_VCPKG_REGISTRIES_CACHE to be consistent with all other input paths. This means there is a new intentionally-undocumented-for-now command line switch x-registries-cache .
* Fixes VS Code to create new files with LF extensions by default.