Teach the powershell one-liner installer to work on Linux. (#611)
Resolves https://github.com/microsoft/vcpkg/issues/25487

21: Fix $IsWindows for PowerShell 5.x and earlier
64: If not on Windows, we need to make our downloads executable
82: Only Windows ends in exe.
118: Hook up various downloads.

Note that we don't do SHA checks here like we do in bootstrap-vcpkg.sh; this is consistent with vcpkg-init (the shell script). Unfortunately this is a result of the vcpkg binary needing to contain the SHA of the tarball containing the support files, including vcpkg-init. (We would need to form a SHA to itself)