Reorder VcpkgPaths.
Delete the details:: namespace here which saves us a whole indent for this file.
Reorder all functions in the .cpp to be in the same order they are declared in the header.
Reorder several functions marked static to be within the unnamed namespace, and drop static.

No function bodies are edited.