Remove the ee-ts dependency.
Fixes signed builds. The previous prepare-deploy was attempting to remove what it thought was a phantom dependency on the typescript compiler from the deployed output. This failed because the DevDiv build lab pushed out changes that blocked the unrestrained dependency on rimraf used by that script. This change drops the dependency on ee-ts which created the need to include the typescript compiler in the first place.

Also fixed test failures that ee-ts was apparently hiding?!