Artifacts: Get rid of the markdown parser.
We were always just emitting the markdown output to the console. This feature was very expensive, both in terms of deployment size and in terms of execution time.

When running `vcpkg find artifact gcc`, this change reduces the execution time from ~708 ms to ~378 ms, or 1.8 times as fast. It also cuts the deployed size of the ecmascript by 48%. (This is very close to letting us just embed the ecmascript directly which means we can avoid an extra download when someone touches artifacts)