Disable metrics when the user is not opted in to Visual Studio telemetry when vcpkg is deployed with Visual Studio.
This required moving the bundleconfig to "next to the .exe" rather than "within vcpkg_root", and corresponding test changes, since we need to know information out of the bundle before we know where VCPKG_ROOT is (and maybe for commands that don't need VCPKG_ROOT in the first place).

This also changes `--no-disablemetrics` to force enable metrics if the user does that for some reason, even if there's a file named vcpkg.disable-metrics.

Drive bys:
* test_registry_key and get_registry_string where leaktrocities.
* get_registry_string was racy.
* get_registry_string accepted REG_MULTI_SZ despite that not actually being a string type. (It's vector<string>)