Remove the vcpkg-find-acquire-program test port.
Hopefully resolves test failures in https://github.com/microsoft/vcpkg-tool/pull/1129 -- asset caching is not turned on in the tool repo so testing things that need lots of external servers is problematic.

This was added back when the tool and registry were in the same repo in attempt to validate *registry* changes, not tool changes. As a result it doesn't really belong here.

The one thing that might also have been tested here is handling of old CONTROL files ports so I added an explicit test port for that.