Temporarily disable GitHub telemetry. (#1131)
The vcpkg team is interested in having an idea of how many vcpkg users there are. Unfortunately, ephemeral CI systems will make a new machine every build or something like that, which can make one user look like hundreds of users. To help reduce that problem, in https://github.com/microsoft/vcpkg-tool/pull/1113 @dan-shaw added reporting of GitHub IDs so that we may deduplicate all CI runs for a project as one 'user'. Unfortunately, there was a miscommunication and privacy folks want to look more closely to make sure that nothing about these IDs might be PII. (We only care about the number of users, not who those users are.)

Our understanding is that these IDs are in fact not PII, and after confirmation we'll turn this back on.