Don't open multiple handles to files to check for absolute paths.
Rather than opening separate file handles, and reading the first part with one and the rest with the other, introduce best_effort_read_contents_if_shebang which implements this test directly when doing the first read buffer.

Also fixes read_contents to avoid needing to copy the entire file down by 3 bytes if the file starts with a BOM using the same idea.