Make print_plan actually print.
print_plan does 2 things. First, it formats the action plan and prints it to the console. Second, if there were removals in the plan and the intent is not recursive, it terminates vcpkg. A function named 'print' should not be terminating vcpkg.

Moreover, only one caller actually desires this terminating behavior: the install command. All other users hard coded a value of true.

This suggests that the decision to terminate should be localized to that one caller rather than spread over everything.