Remove unnecessary use of std::array
std::array is actually not that cheap to compile and we did not really need it in these cases. ( https://raw.githubusercontent.com/ned14/stl-header-heft/master/graphs/msvs-2019.png )

The remaining place we still use it is in tools.cpp where it represents a version breakdown; that change was invasive so I've separated this off for now.