Add switch to force classic mode. (#1535)
A customer wanted to run vcpkg in a place where they could not control that a `vcpkg.json` was in a directory above, and requested a way to force classic mode. One used to be able to do this with the `--no-manifest` feature flag, but that was removed some years ago.

This adds a new switch, `--classic`, which effectively skips looking for a manifest.