Respect and emit the correct encoding for serialized JSON files
When vcpkg updates the vcpkg.json file, it was aggressively and unconditionally overwriting the file with LF line endings.

This can cause issues with git repos that use the "core.autocrlf=true" option on Windows, as well as with any text processing tools which might wish to impose CRLF line endings (just as vcpkg.exe was imposing LF), resulting in encoding tugs of war.

In my particular case, with "autocrlf=true", a vcpkg.json file that's checked out as part of a repo will get modified on Windows from CRLF to LF line endings, which causes the file to appear modified when it shouldn't. Basically, vcpkg and git get to disagree on the correct encoding of the file and engage in such a tug of war.

It is understood that there are differing opinions regarding this setting, but vcpkg should remain as friendly as possible regardless of the setting used by some repo or user.

This is what it looked like in my repo after vcpkg touched the vcpkg.json file:
```
C:\repo>git status
On branch show-and-tell
Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
        modified:   vcpkg.json

C:\repo>git diff vcpkg.json
warning: in the working copy of 'vcpkg.json', LF will be replaced by CRLF the next time Git touches it
```

When I went in this codebase I expected to do one of:
- Add line encoding awareness to JSON serialization, so updated files would be rewritten using the original encoding.
- If that turned out too complicated, maybe add an indirection to serialization to avoid writing the file if the contents end up unchanged.

What I found is that the first option was already mostly implemented. The JSON parser already recognizes and reports the newline encodings of the file it parses. But this encoding is then left unused and forgotten.

This change aims to close that loop. Here is what's been done:
- Add a new `Json::ParsedObject` structure to hold a `Json::Object` together with its `Json::JsonStyle`. `Json::parse_object()` now returns new object.
  - Modify code throughout to account for the resulting indirection.
  - Transfer the newline style from parser to serializer wherever possible.
  - Fix the `Json::stringify()` functions that get a newline style to actually use that style.
- Modify the default newline encoding to match the platform.
  - This might be controversial. For any **new** files created by vcpkg, arguably they should be created using the platform's natural newline encoding.
  - Explicitly request a style of LF for calls to `Json::stringify` throughout, to keep the same behavior as before wherever I was unsure of the right thing to do. These should probably be reviewed by someone with better expertise in vcpkg use cases.
- Change test comparison of multiline strings to use the already-existing `REQUIRE_LINES()` macro.
  - Remove the delimiter parameter of `vcpkg::Strings::split_keep_empty` and rename it to `split_lines_keep_empty`, so that line splitting can now recognize the multiple line endings in circulation. Note that this is only used to implement the `REQUIRE_LINES()` macro for use in tests.