Address review feedback and fix an unrelated end-to-end test issue.
The test issue was found by running he end-to-end tests locally, to validate that the changes will resolve the issues foundby the PR validation.

One of the tests needs to use the 7z tool to unzip a ninja package. Added the necessary entries to the embedded JSON to make the tool available, thank you for the insight, @BillyONeal.