[spdx] Add installed package files to SPDX SBOM file (#1744)
* [spdx] Add installed package files to SPDX SBOM file

Also remove repetition of relationships. These are no longer present in v3 of
the SPDX spec. Keeping these relationships would also make the SBOM very
noisy.

* Add AbiTagSbomInfo to track SBOM changes

Because this is changing the output packages, it needs to change the output
package ABI (so we'll regenerate all packages with the new SBOM information)