Get props and targets from the scripts bundle.
I'm not positive exactly what I was thinking in https://github.com/microsoft/vcpkg-tool/pull/976 ; probably though we'd be moving the authoritative place to the tool repo. But the result of that has been that the VS copy is shipping very old bits for a while.

Bug discvoered by @vicroms .