Include NuGet console output on failures. (#1800)
Prior to this we were dropping NuGet console output on the floor requiring folks to use --debug.

There is a behavior change here in that if the exit code was 0 before but the console output included "Authentication may require manual action." we would print a warning anyway, now the warning is only printed on failure.

Drive by: Audit for 'Nuget' -> 'NuGet' for correct branding.