Allow MSVC on non Windows platforms for cross compilation
This allows MSVC to be used on Linux. I tested this with a setup using
the msvc-wine project that provided wrapper scripts around the Wine
compatibility layer to run the MSVC compiler on other platforms. By
forcing an external toolset this forces vcpkg to use the default
compiler or the CC/CXX environment flags, which with use of I was able
to cross compile some projects from Linux to Windows/MSVC.