Fix GitHub API endpoint selection based on host parsing
Previously, any provided GitHub URL was treated as a GitHub Enterprise
Server (GHES) instance when selecting the API endpoint. This assumption was
incorrect because the URL can also be for GitHub.com (e.g., https://github.com),
which led to using the wrong API endpoint.

This commit adds a function to extract the host from a URL. This is used to
differentiate between GitHub.com and GHES instances to select the correct
API endpoint.